Revert "Added Inspector client code, including data acquisition and sending of op, task and attributes. (#684)"
This reverts commit af60d8ffaae135869dcc3fc33aec74fe7ee1c83e.
