+++ CMakeLists.txt
@@ -32,7 +32,7 @@ if (NOT WEB OR EMSCRIPTEN_PTHREADS)
 endif ()
 
 if (CMAKE_BUILD_TYPE MATCHES "Debug")
-    option(TGFX_BUILD_INSPECTOR "Enable profiling support for the TGFX Inspector tool" ON)
+    option(TGFX_USE_INSPECTOR "Enable profiling support for the TGFX Inspector tool" ON)
 endif ()
 
 if (APPLE OR (WEB AND NOT TGFX_USE_THREADS))
@@ -103,7 +103,7 @@ message("TGFX_USE_WEBP_DECODE: ${TGFX_USE_WEBP_DECODE}")
 message("TGFX_USE_WEBP_ENCODE: ${TGFX_USE_WEBP_ENCODE}")
 message("TGFX_BUILD_TESTS: ${TGFX_BUILD_TESTS}")
 message("TGFX_USE_ASYNC_PROMISE: ${TGFX_USE_ASYNC_PROMISE}")
-message("TGFX_BUILD_INSPECTOR: ${TGFX_BUILD_INSPECTOR}")
+message("TGFX_USE_INSPECTOR: ${TGFX_USE_INSPECTOR}")
 
 if (NOT CMAKE_OSX_DEPLOYMENT_TARGET)
     if (DEPLOYMENT_TARGET)
@@ -150,18 +150,6 @@ if (TGFX_BUILD_LAYERS)
             src/layers/*.*)
     list(APPEND TGFX_FILES ${LAYERS_FILES})
 endif ()
-
-if (TGFX_BUILD_INSPECTOR)
-    set(LZ4_DIR third_party/lz4/lib)
-    list(APPEND LZ4File ${LZ4_DIR}/lz4.c)
-    list(APPEND TGFX_INCLUDES third_party/lz4/lib)
-    file(GLOB_RECURSE INSPECTOR_FILES src/debug/*.*)
-    file(GLOB_RECURSE SERIALIZATION_FILE src/serialization/*.*)
-    list(APPEND TGFX_FILES ${SERIALIZATION_FILE} ${INSPECTOR_FILES} ${LZ4File})
-    list(APPEND TGFX_DEFINES TGFX_BUILD_INSPECTOR)
-    list(APPEND TGFX_INCLUDES third_party/flatbuffers/include)
-    list(APPEND TGFX_STATIC_VENDORS flatbuffers)
-endif()
 
 if (CMAKE_CXX_COMPILER_ID MATCHES "Clang")
     list(APPEND TGFX_COMPILE_OPTIONS -Werror -Wall -Wextra -Weffc++ -Wconversion -pedantic -Werror=return-type -Wno-unused-command-line-argument)
@@ -298,6 +286,18 @@ elseif (APPLE)
 endif ()
 
 file(GLOB_RECURSE EGL_PLATFORM_FILES src/gpu/opengl/egl/*.*)
+
+if (TGFX_USE_INSPECTOR)
+    set(INSPECTOR_DIR ${CMAKE_CURRENT_LIST_DIR}/inspector)
+    list(APPEND TGFX_INCLUDES ${INSPECTOR_DIR}/client/include ${INSPECTOR_DIR}/client/src ${INSPECTOR_DIR}/common)
+    list(APPEND TGFX_STATIC_VENDORS inspector)
+
+    list(APPEND TGFX_DEFINES TGFX_USE_INSPECTOR)
+    list(APPEND TGFX_INCLUDES third_party/flatbuffers/include)
+    file(GLOB_RECURSE SERIALIZATION_FILE src/serialization/*.*)
+    list(APPEND TGFX_FILES ${SERIALIZATION_FILE})
+    list(APPEND TGFX_STATIC_VENDORS flatbuffers)
+endif ()
 
 if (TGFX_USE_QT)
     set(CMAKE_AUTOMOC ON)
