+++ DEPS
@@ -79,11 +79,6 @@
         "url": "https://github.com/google/highway.git",
         "commit": "a523516d35e22a4ba8e2e70a319062cb87352de6",
         "dir": "third_party/highway"
-      },
-      {
-        "url": "https://github.com/lz4/lz4.git",
-        "commit": "cacca37747572717ceb1f156eb9840644205ca4f",
-        "dir": "third_party/lz4"
       }
     ]
   },
