+++ inspector/client/include/Define.h
@@ -15,13 +15,9 @@
 //  and limitations under the license.
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
-
 #pragma once
-#include <chrono>
-#include <cstdint>
-namespace inspector {
-template <typename T>
-int64_t GetCurrentTime() {
-  return std::chrono::duration_cast<T>(std::chrono::system_clock::now().time_since_epoch()).count();
-}
-}  // namespace inspector
+
+#include "LayerProfiler.h"
+
+#define SEND_LAYER_DATA(data) inspector::LayerProfiler::Get().setData(data)
+#define LAYER_CALLBACK(func) inspector::LayerProfiler::Get().setCallBack(func)
