+++ inspector/client/src/LayerProfiler.cpp
@@ -2,7 +2,7 @@
 //
 //  Tencent is pleased to support the open source community by making tgfx available.
 //
-//  Copyright (C) 2025 Tencent. All rights reserved.
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
 //
 //  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
 //  in compliance with the License. You may obtain a copy of the License at
@@ -17,11 +17,11 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 #include "LayerProfiler.h"
 #include <chrono>
-#include <cstring>
 #include <thread>
+#include <cstring>
 #include "ProcessUtils.h"
+#include "TimeUtils.h"
 #include "Protocol.h"
-#include "tgfx/core/Clock.h"
 
 namespace inspector {
 
@@ -30,25 +30,25 @@ static const char* addr = "255.255.255.255";
 static uint16_t broadcastPort = 8086;
 #endif
 
-LayerProfiler::LayerProfiler() {
+LayerProfiler::LayerProfiler(){
 #ifndef __EMSCRIPTEN__
   listenSocket = std::make_shared<ListenSocket>();
   for (uint16_t i = 0; i < broadcastNum; i++) {
     broadcasts[i] = std::make_shared<UdpBroadcast>();
-    isUDPOpened = isUDPOpened && broadcasts[i]->openConnect(addr, broadcastPort + i);
+    isUDPOpened = isUDPOpened && broadcasts[i]->open(addr, broadcastPort + i);
   }
 
 #endif
-  epoch = tgfx::Clock::Now<std::chrono::seconds>();
+  epoch = GetCurrentTime<std::chrono::seconds>();
   spawnWorkTread();
 }
 
 LayerProfiler::~LayerProfiler() {
   stopFlag.store(true, std::memory_order_release);
-  if (sendThread && sendThread->joinable()) {
+  if(sendThread && sendThread->joinable()) {
     sendThread->join();
   }
-  if (recvThread && recvThread->joinable()) {
+  if(recvThread && recvThread->joinable()) {
     recvThread->join();
   }
 }
@@ -61,11 +61,11 @@ void LayerProfiler::sendWork() {
   const auto procname = GetProcessName();
   const auto pnsz = std::min<size_t>(strlen(procname), WelcomeMessageProgramNameSize - 1);
   uint16_t port = TCPPortProvider::Get().getValidPort();
-  if (!listenSocket->listenSock(port, 4)) {
+  if (!listenSocket->listen(port, 4)) {
     return;
   }
   size_t broadcastLen = 0;
-  auto broadcastMsg = GetBroadcastMessage(procname, pnsz, broadcastLen, port, ToolType::LayerTree);
+  auto broadcastMsg = GetBroadcastMessage(procname, pnsz, broadcastLen, port, static_cast<uint8_t>(MsgType::LayerTree));
   long long lastBroadcast = 0;
   while (!stopFlag.load(std::memory_order_acquire)) {
     while (!stopFlag.load(std::memory_order_acquire)) {
@@ -75,13 +75,13 @@ void LayerProfiler::sendWork() {
         lastBroadcast = t;
         for (uint16_t i = 0; i < broadcastNum; i++) {
           if (broadcasts[i]) {
-            const auto ts = tgfx::Clock::Now<std::chrono::seconds>();
+            const auto ts = GetCurrentTime<std::chrono::seconds>();
             broadcastMsg.activeTime = static_cast<int32_t>(ts - epoch);
-            broadcasts[i]->sendData(broadcastPort + i, &broadcastMsg, broadcastLen);
+            broadcasts[i]->send(broadcastPort + i, &broadcastMsg, broadcastLen);
           }
         }
       }
-      socket = listenSocket->acceptSock();
+      socket = listenSocket->accept();
       if (socket) {
         break;
       }
@@ -94,10 +94,10 @@ void LayerProfiler::sendWork() {
       }
       if (queue.size_approx() != 0) {
         std::vector<uint8_t> data;
-        if (queue.try_dequeue(data)) {
+        if(queue.try_dequeue(data)) {
           int size = (int)data.size();
-          socket->sendData(&size, sizeof(int));
-          socket->sendData(data.data(), data.size());
+          socket->send(&size, sizeof(int));
+          socket->send(data.data(), data.size());
         }
       }
     }
@@ -139,6 +139,6 @@ void LayerProfiler::setData(const std::vector<uint8_t>& data) {
   queue.enqueue(data);
 }
 void LayerProfiler::setCallBack(std::function<void(const std::vector<uint8_t>&)> callback) {
-  this->callback = std::move(callback);
+    this->callback = std::move(callback);
 }
 }  // namespace inspector
