+++ inspector/client/src/LayerProfiler.h
@@ -2,7 +2,7 @@
 //
 //  Tencent is pleased to support the open source community by making tgfx available.
 //
-//  Copyright (C) 2025 Tencent. All rights reserved.
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
 //
 //  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
 //  in compliance with the License. You may obtain a copy of the License at
@@ -20,21 +20,21 @@
 #include <queue>
 #include <thread>
 #include <vector>
+#include "concurrentqueue.h"
 #include "Protocol.h"
 #include "Socket.h"
 #include "TCPPortProvider.h"
-#include "concurrentqueue.h"
 namespace inspector {
 class LayerProfiler {
  public:
   static LayerProfiler& Get() {
-    static LayerProfiler instance;
-    return instance;
+      static LayerProfiler instance;
+      return instance;
   }
-  LayerProfiler(const LayerProfiler&) = delete;
-  LayerProfiler(LayerProfiler&&) = delete;
-  LayerProfiler& operator=(const LayerProfiler&) = delete;
-  LayerProfiler& operator=(LayerProfiler&&) = delete;
+    LayerProfiler(const LayerProfiler&) = delete;
+    LayerProfiler(LayerProfiler&&) = delete;
+    LayerProfiler& operator=(const LayerProfiler&) = delete;
+    LayerProfiler& operator=(LayerProfiler&&) = delete;
   ~LayerProfiler();
   void setData(const std::vector<uint8_t>& data);
   void setCallBack(std::function<void(const std::vector<uint8_t>&)> callback);
