+++ inspector/client/src/TCPPortProvider.cpp
@@ -2,7 +2,7 @@
 //
 //  Tencent is pleased to support the open source community by making tgfx available.
 //
-//  Copyright (C) 2025 Tencent. All rights reserved.
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
 //
 //  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
 //  in compliance with the License. You may obtain a copy of the License at
@@ -37,15 +37,14 @@ uint16_t TCPPortProvider::getValidPort() {
   uint16_t randomPort = dis(gen);
   std::unique_ptr<ListenSocket> listenSocket = std::make_unique<ListenSocket>();
   while (usedPortSet.find(randomPort) != usedPortSet.end() ||
-         !listenSocket->listenSock(randomPort, 4)) {
+         !listenSocket->listen(randomPort, 4)) {
     randomPort = dis(gen);
   }
   usedPortSet.insert(randomPort);
   return randomPort;
 }
-
 bool TCPPortProvider::clearUsedPort(uint16_t port) {
-  if (usedPortSet.find(port) != usedPortSet.end()) {
+  if(usedPortSet.find(port) != usedPortSet.end()) {
     usedPortSet.erase(port);
     return true;
   }
