+++ inspector/client/src/TCPPortProvider.h
@@ -2,7 +2,7 @@
 //
 //  Tencent is pleased to support the open source community by making tgfx available.
 //
-//  Copyright (C) 2025 Tencent. All rights reserved.
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
 //
 //  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
 //  in compliance with the License. You may obtain a copy of the License at
@@ -20,7 +20,7 @@
 #include <set>
 namespace inspector {
 class TCPPortProvider {
- public:
+public:
   static TCPPortProvider& Get() {
     static TCPPortProvider instance;
     return instance;
@@ -31,9 +31,9 @@ class TCPPortProvider {
   TCPPortProvider& operator=(TCPPortProvider&& provider) = delete;
   uint16_t getValidPort();
   bool clearUsedPort(uint16_t port);
-
- private:
+private:
   TCPPortProvider() = default;
   std::set<uint16_t> usedPortSet = {};
 };
-}  // namespace inspector
+}
+
