+++ inspector/common/LayerInspectorProtocol.h
@@ -15,23 +15,24 @@
 //  and limitations under the license.
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
+
 #pragma once
-
-#include "Inspector.h"
-#include "LayerProfiler.h"
-#include "Scoped.h"
-
-#define SEND_LAYER_DATA(data) inspector::LayerProfiler::Get().setData(data)
-#define LAYER_CALLBACK(func) inspector::LayerProfiler::Get().setCallBack(func)
-
-#define FrameMark inspector::Inspector::SendFrameMark(nullptr)
-
-#define ScopedMark(type, active) inspector::Scoped scoped(type, active)
-#define OperateMark(type) ScopedMark(type, true)
-#define TaskMark(type) ScopedMark(type, true)
-
-#define AttributeName(name, value) inspector::Inspector::SendAttributeData(name, value)
-#define AttributeNameEnum(name, value, type)                                 \
-  inspector::Inspector::SendAttributeData(name, static_cast<uint8_t>(value), \
-                                          static_cast<uint8_t>(type))
-#define AttributeEnum(value, type) AttributeNameEnum(#value, value, type)
+#include <cstdint>
+namespace inspector {
+  enum class LayerInspectorMsgType : uint8_t {
+    EnableLayerInspector,
+    HoverLayerAddress,
+    SelectedLayerAddress,
+    SerializeAttribute,
+    SerializeSubAttribute,
+    FlushAttribute,
+    FlushLayerTree,
+    FlushImage,
+    PickedLayerAddress,
+    FlushAttributeAck,
+    LayerTree,
+    LayerAttribute,
+    LayerSubAttribute,
+    ImageData
+  };
+}  // namespace tgfx
