+++ inspector/common/ProcessUtils.h
@@ -2,7 +2,7 @@
 //
 //  Tencent is pleased to support the open source community by making tgfx available.
 //
-//  Copyright (C) 2025 Tencent. All rights reserved.
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
 //
 //  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
 //  in compliance with the License. You may obtain a copy of the License at
@@ -18,21 +18,11 @@
 
 #pragma once
 #include <cstdint>
+#include "Protocol.h"
+
 namespace inspector {
-enum class LayerInspectorMsgType : uint8_t {
-  EnableLayerInspector,
-  HoverLayerAddress,
-  SelectedLayerAddress,
-  SerializeAttribute,
-  SerializeSubAttribute,
-  FlushAttribute,
-  FlushLayerTree,
-  FlushImage,
-  PickedLayerAddress,
-  FlushAttributeAck,
-  LayerTree,
-  LayerAttribute,
-  LayerSubAttribute,
-  ImageData
-};
+uint64_t GetPid();
+const char* GetProcessName();
+BroadcastMessage GetBroadcastMessage(const char* procname, size_t pnsz, size_t& len, uint16_t port,
+                                     uint8_t type);
 }  // namespace inspector
