+++ inspector/common/Protocol.h
@@ -2,7 +2,7 @@
 //
 //  Tencent is pleased to support the open source community by making tgfx available.
 //
-//  Copyright (C) 2025 Tencent. All rights reserved.
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
 //
 //  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
 //  in compliance with the License. You may obtain a copy of the License at
@@ -17,40 +17,23 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
-#include "Inspector.h"
-#include "tgfx/core/Clock.h"
-
 namespace inspector {
-class Scoped {
 
- public:
-  Scoped(const Scoped&) = delete;
-  Scoped(Scoped&&) = delete;
-  Scoped& operator=(const Scoped&) = delete;
-  Scoped& operator=(Scoped&&) = delete;
+inline constexpr int broadcastNum = 5;
 
-  Scoped(OpTaskType type, bool isActive) : active(isActive), type(type) {
-    if (!active) {
-      return;
-    }
-    MsgPrepare(MsgType::OperateBegin);
-    MemWrite(&item.operateBegin.nsTime, tgfx::Clock::Now<std::chrono::nanoseconds>());
-    MemWrite(&item.operateBegin.type, type);
-    MsgCommit();
-  }
+inline constexpr uint8_t WelcomeMessageProgramNameSize = 64;
+inline constexpr uint8_t ProtocolVersion = 1;
+inline constexpr uint16_t BroadcastVersion = 1;
 
-  ~Scoped() {
-    if (!active) {
-      return;
-    }
-    MsgPrepare(MsgType::OperateEnd);
-    MemWrite(&item.operateEnd.nsTime, tgfx::Clock::Now<std::chrono::nanoseconds>());
-    MemWrite(&item.operateEnd.type, type);
-    MsgCommit();
-  }
+enum class MsgType : uint8_t { FrameCapture = 0, LayerTree = 1 };
 
- private:
-  bool active = false;
-  OpTaskType type = OpTaskType::Unknown;
+struct BroadcastMessage {
+  uint8_t type;
+  uint16_t listenPort;
+  uint32_t protocolVersion;
+  uint64_t pid;
+  int32_t activeTime;  // in seconds
+  char programName[WelcomeMessageProgramNameSize];
 };
+
 }  // namespace inspector
