+++ inspector/common/TimeUtils.h
@@ -18,11 +18,12 @@
 
 #pragma once
 #include <cstdint>
-#include "Protocol.h"
-
+#include <chrono>
 namespace inspector {
-uint64_t GetPid();
-const char* GetProcessName();
-BroadcastMessage GetBroadcastMessage(const char* procname, size_t pnsz, size_t& len, uint16_t port,
-                                     ToolType type);
-}  // namespace inspector
+template<typename T>
+int64_t GetCurrentTime() {
+  return std::chrono::duration_cast<T>(
+                              std::chrono::system_clock::now().time_since_epoch())
+                              .count();
+}
+}  // namespace tgfx
