+++ src/gpu/DrawingManager.cpp
@@ -21,7 +21,6 @@
 #include "ProxyProvider.h"
 #include "core/AtlasCellDecodeTask.h"
 #include "core/AtlasManager.h"
-#include "core/utils/Profiling.h"
 #include "gpu/proxies/RenderTargetProxy.h"
 #include "gpu/proxies/TextureProxy.h"
 #include "gpu/tasks/GenerateMipmapsTask.h"
@@ -160,7 +159,6 @@ void DrawingManager::addSemaphoreWaitTask(std::shared_ptr<Semaphore> semaphore) 
 }
 
 std::shared_ptr<CommandBuffer> DrawingManager::flush(BackendSemaphore* signalSemaphore) {
-  TaskMark(inspector::OpTaskType::Flush);
   while (!compositors.empty()) {
     auto compositor = compositors.back();
     // The makeClosed() method may add more compositors to the list.
@@ -175,23 +173,17 @@ std::shared_ptr<CommandBuffer> DrawingManager::flush(BackendSemaphore* signalSem
     clearAtlasCellCodecTasks();
     return nullptr;
   }
-  {
-    TaskMark(inspector::OpTaskType::ResourceTask);
-    for (auto& task : resourceTasks) {
-      task->execute(context);
-      task = nullptr;
-    }
+  for (auto& task : resourceTasks) {
+    task->execute(context);
+    task = nullptr;
   }
   uploadAtlasToGPU();
   resourceTasks.clear();
   proxyProvider->clearSharedVertexBuffer();
   auto commandEncoder = context->gpu()->createCommandEncoder();
-  {
-    TaskMark(inspector::OpTaskType::RenderTask);
-    for (auto& task : renderTasks) {
-      task->execute(commandEncoder.get());
-      task = nullptr;
-    }
+  for (auto& task : renderTasks) {
+    task->execute(commandEncoder.get());
+    task = nullptr;
   }
   renderTasks.clear();
   if (signalSemaphore != nullptr) {
