+++ src/gpu/Window.cpp
@@ -18,7 +18,6 @@
 
 #include "tgfx/gpu/Window.h"
 #include "core/utils/Log.h"
-#include "core/utils/Profiling.h"
 #include "tgfx/gpu/Device.h"
 
 namespace tgfx {
@@ -58,7 +57,6 @@ void Window::freeSurface() {
 }
 
 void Window::present(Context* context, int64_t presentationTime) {
-  FrameMark;
   std::lock_guard<std::mutex> autoLock(locker);
   if (!checkContext(context)) {
     return;
