+++ src/gpu/ops/ShapeDrawOp.cpp
@@ -18,8 +18,6 @@
 
 #include "ShapeDrawOp.h"
 #include "core/PathTriangulator.h"
-#include "core/utils/Log.h"
-#include "core/utils/Profiling.h"
 #include "gpu/ProxyProvider.h"
 #include "gpu/Quad.h"
 #include "gpu/RectsVertexProvider.h"
@@ -51,11 +49,6 @@ ShapeDrawOp::ShapeDrawOp(std::shared_ptr<GPUShapeProxy> proxy, Color color, cons
 }
 
 void ShapeDrawOp::execute(RenderPass* renderPass) {
-  OperateMark(inspector::OpTaskType::ShapeDrawOp);
-  AttributeTGFXName("color", color);
-  AttributeTGFXName("uvMatrix", uvMatrix);
-  AttributeNameEnum("blenderMode", getBlendMode(), inspector::CustomEnumType::BlendMode);
-  AttributeNameEnum("aaType", getAAType(), inspector::CustomEnumType::AAType);
   if (shapeProxy == nullptr) {
     return;
   }
