+++ src/gpu/tasks/TextureUploadTask.cpp
@@ -17,7 +17,6 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include "TextureUploadTask.h"
-#include "core/utils/Profiling.h"
 #include "gpu/Texture.h"
 
 namespace tgfx {
@@ -28,8 +27,6 @@ TextureUploadTask::TextureUploadTask(std::shared_ptr<ResourceProxy> proxy,
 }
 
 std::shared_ptr<Resource> TextureUploadTask::onMakeResource(Context* context) {
-  TaskMark(inspector::OpTaskType::TextureUploadTask);
-  AttributeName("mipmaped", mipmapped);
   if (source == nullptr) {
     return nullptr;
   }
