+++ src/layers/DisplayList.cpp
@@ -24,8 +24,8 @@
 #include "layers/RootLayer.h"
 #include "layers/TileCache.h"
 
-#ifdef TGFX_BUILD_INSPECTOR
-#include "debug/LayerViewerManager.h"
+#ifdef TGFX_USE_INSPECTOR
+#include "layers/LayerViewerManager.h"
 #endif
 
 namespace tgfx {
@@ -116,7 +116,7 @@ static int64_t ChangeZoomScalePrecision(int64_t zoomScaleInt, int oldPrecision, 
 
 DisplayList::DisplayList() : _root(RootLayer::Make()) {
   _root->_root = _root.get();
-#ifdef TGFX_BUILD_INSPECTOR
+#ifdef TGFX_USE_INSPECTOR
   auto& layerInspectorManager = LayerViewerManager::Get();
   layerInspectorManager.setDisplayList(this);
 #endif
@@ -243,7 +243,7 @@ void DisplayList::render(Surface* surface, bool autoClear) {
   if (!surface) {
     return;
   }
-#ifdef TGFX_BUILD_INSPECTOR
+#ifdef TGFX_USE_INSPECTOR
   LayerViewerManager::Get().RenderImageAndSend(surface->getContext());
 #endif
   _hasContentChanged = false;
