+++ src/layers/LayerViewerManager.cpp
@@ -2,7 +2,7 @@
 //
 //  Tencent is pleased to support the open source community by making tgfx available.
 //
-//  Copyright (C) 2025 Tencent. All rights reserved.
+//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
 //
 //  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
 //  in compliance with the License. You may obtain a copy of the License at
@@ -15,6 +15,7 @@
 //  and limitations under the license.
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
+#ifdef TGFX_USE_INSPECTOR
 #include "LayerViewerManager.h"
 #include <functional>
 #include <string>
@@ -214,3 +215,4 @@ void LayerViewerManager::AddHighLightOverlay(Color color, std::shared_ptr<Layer>
   highLightLayerIndex = hoverdLayer->getChildIndex(highlightLayer);
 }
 }  // namespace tgfx
+#endif
