+++ src/serialization/ColorSerialization.cpp
@@ -16,6 +16,8 @@
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
+#ifdef TGFX_USE_INSPECTOR
+
 #include "ColorSerialization.h"
 
 namespace tgfx {
@@ -38,3 +40,4 @@ std::shared_ptr<Data> ColorSerialization::Serialize(const Color* color) {
   return Data::MakeWithCopy(fbb.GetBuffer().data(), fbb.GetBuffer().size());
 }
 }  // namespace tgfx
+#endif
