+++ src/serialization/LayerSerialization.cpp
@@ -15,6 +15,7 @@
 //  and limitations under the license.
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
+#ifdef TGFX_USE_INSPECTOR
 
 #include "LayerSerialization.h"
 #include <tgfx/layers/ImageLayer.h>
@@ -168,3 +169,4 @@ void LayerSerialization::SerializeBasicLayerImpl(flexbuffers::Builder& fbb, cons
 }
 
 }  // namespace tgfx
+#endif
