+++ src/serialization/LayerStyleSerialization.cpp
@@ -15,6 +15,7 @@
 //  and limitations under the license.
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
+#ifdef TGFX_USE_INSPECTOR
 
 #include "LayerStyleSerialization.h"
 #include <tgfx/layers/filters/DropShadowFilter.h>
@@ -107,3 +108,4 @@ std::shared_ptr<Data> LayerStyleSerialization::Serialize(const LayerStyle* layer
   return Data::MakeWithCopy(fbb.GetBuffer().data(), fbb.GetBuffer().size());
 }
 }  // namespace tgfx
+#endif
