+++ src/serialization/MatrixSerialization.cpp
@@ -15,6 +15,7 @@
 //  and limitations under the license.
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
+#ifdef TGFX_USE_INSPECTOR
 
 #include "MatrixSerialization.h"
 #include <string>
@@ -42,3 +43,4 @@ std::shared_ptr<Data> MatrixSerialization::Serialize(const Matrix* matrix) {
   return Data::MakeWithCopy(fbb.GetBuffer().data(), fbb.GetBuffer().size());
 }
 }  // namespace tgfx
+#endif
