+++ src/serialization/RectSerialization.cpp
@@ -15,6 +15,7 @@
 //  and limitations under the license.
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
+#ifdef TGFX_USE_INSPECTOR
 
 #include "RectSerialization.h"
 
@@ -39,3 +40,4 @@ std::shared_ptr<Data> RectSerialization::Serialize(const Rect* rect) {
   return Data::MakeWithCopy(fbb.GetBuffer().data(), fbb.GetBuffer().size());
 }
 }  // namespace tgfx
+#endif
