+++ src/serialization/SerilaizationUtils.cpp
@@ -16,6 +16,7 @@
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
+#ifdef TGFX_USE_INSPECTOR
 #include "ColorSerialization.h"
 #include "LayerFilterSerialization.h"
 #include "LayerSerialization.h"
@@ -494,3 +495,4 @@ void SerializeUtils::SetFlexBufferMap(flexbuffers::Builder& fbb, const char* key
   });
 }
 }  // namespace tgfx
+#endif
