+++ vendor.json
@@ -251,6 +251,15 @@
       }
     },
     {
+      "name": "inspector",
+      "dir": "../inspector",
+      "cmake": {
+        "targets": [
+          "inspector"
+        ]
+      }
+    },
+    {
       "name": "highway",
       "cmake": {
         "targets": [
